requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
notion-client==2.1.0
openai==1.35.7
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyPDF2==3.0.1
rich==13.7.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
bs4==0.0.1
azure-ai-documentintelligence==1.0.0b4